[Security Solution] [Cases] Follow up jest tests for #84587 (#86231)
