+++ x-pack/plugins/case/server/routes/api/__mocks__/request_responses.ts
@@ -3,9 +3,15 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
-import { CasePostRequest, CasesConfigureRequest, ConnectorTypes } from '../../../../common/api';
+import {
+  CasePostRequest,
+  CasesConfigureRequest,
+  ConnectorTypes,
+  PostPushRequest,
+} from '../../../../common/api';
 // eslint-disable-next-line @kbn/eslint/no-restricted-paths
 import { FindActionResult } from '../../../../../actions/server/types';
+import { params } from '../cases/configure/mock';
 
 export const newCase: CasePostRequest = {
   title: 'My new case',
@@ -75,4 +81,20 @@ export const newConfiguration: CasesConfigureRequest = {
     fields: null,
   },
   closure_type: 'close-by-pushing',
+};
+
+export const newPostPushRequest: PostPushRequest = {
+  params: params[ConnectorTypes.jira],
+  connector_type: ConnectorTypes.jira,
+};
+
+export const executePushResponse = {
+  status: 'ok',
+  data: {
+    title: 'RJ2-200',
+    id: '10663',
+    pushedDate: '2020-12-17T00:32:40.738Z',
+    url: 'https://siem-kibana.atlassian.net/browse/RJ2-200',
+    comments: [],
+  },
 };
