+++ x-pack/plugins/case/server/routes/api/index.ts
@@ -25,7 +25,6 @@ import { initPostCommentApi } from './cases/comments/post_comment';
 
 import { initCaseConfigureGetActionConnector } from './cases/configure/get_connectors';
 import { initGetCaseConfigure } from './cases/configure/get_configure';
-import { initCaseConfigureGetFields } from './cases/configure/get_fields';
 import { initPatchCaseConfigure } from './cases/configure/patch_configure';
 import { initPostCaseConfigure } from './cases/configure/post_configure';
 import { initPostPushToService } from './cases/configure/post_push_to_service';
@@ -54,7 +53,6 @@ export function initCaseApi(deps: RouteDeps) {
   initGetCaseConfigure(deps);
   initPatchCaseConfigure(deps);
   initPostCaseConfigure(deps);
-  initCaseConfigureGetFields(deps);
   initPostPushToService(deps);
   // Reporters
   initGetReportersApi(deps);
